Show live  messages on screen
